src/main/java/com/nesterrovv/vpn/account/mapper/AccountMapper.java
package com.nesterrovv.vpn.account.mapper;

import com.nesterrovv.vpn.account.dto.AccountDto;
import com.nesterrovv.vpn.account.entity.Account;
import org.modelmapper.ModelMapper;
import org.springframework.stereotype.Component;

@Component
public class AccountMapper {

    public AccountDto entityToDto(Account account) {
        ModelMapper modelMapper = new ModelMapper();
        return modelMapper.map(account, AccountDto.class);
    }

    public Account dtoToEntity(AccountDto accountDto) {
        ModelMapper modelMapper = new ModelMapper();
        return modelMapper.map(accountDto, Account.class);
    }

}
